requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==1.4.47
psycopg2-binary==2.9.7
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
